+++ MissionControllerUI/src/main/java/org/ariss/star/Dialogue.java
@@ -1,4 +1,4 @@
-package com.example.hello;
+package org.ariss.star;
 
 import java.io.File;
 import java.io.IOException;
