+++ MissionControllerUI/src/main/java/org/ariss/star/Main.java
@@ -1,10 +1,10 @@
-package com.example.hello;
+package org.ariss.star;
 
 import java.io.IOException;
 
 public class Main {
     public static void main(String [] args)throws IOException{
 
-        HelloApplication.main(args);
+        StarApp.main(args);
     }
 }
