+++ MissionControllerUI/src/main/java/org/ariss/star/RobotEntry.java
@@ -1,4 +1,4 @@
-package com.example.hello;
+package org.ariss.star;
 
 import java.util.Objects;
