--- MissionControllerUI/src/main/resources/com/example/hello/visualize.fxml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Accordion?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Menu?>
-<?import javafx.scene.control.MenuBar?>
-<?import javafx.scene.control.MenuItem?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.control.TitledPane?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.Pane?>
-
-<Pane fx:id="content" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" stylesheets="@vis.css" xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="com.example.hello.visualizeController">
-   <children>
-      <MenuBar>
-        <menus>
-          <Menu mnemonicParsing="false" text="File">
-            <items>
-              <MenuItem mnemonicParsing="false" onAction="#imageChange" text="Add image" />
-            </items>
-            <items>
-              <MenuItem mnemonicParsing="false" onAction="#pathSave" text="Save as" />
-            </items>
-          </Menu>
-            <Menu mnemonicParsing="false" text="Reset Frame">
-               <items>
-                  <MenuItem mnemonicParsing="false" onAction="#reset" text="Reset" />
-               </items>
-            </Menu>
-        </menus>
-      </MenuBar>
-      <ImageView fx:id="imgView" fitHeight="575.0" fitWidth="807.0" layoutX="-6.0" layoutY="25.0" pickOnBounds="true">
-         <image>
-            <Image url="@../images/defaultMarsImage.jpg" />
-         </image>
-      </ImageView>
-      <Accordion layoutX="598.0" layoutY="25.0">
-        <panes>
-          <TitledPane animated="false" text="Enter commands">
-            <content>
-              <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
-                     <children>
-                        <TextArea fx:id="commandBuilder" prefHeight="200.0" prefWidth="200.0" />
-                        <Button fx:id="submit" layoutX="147.0" layoutY="154.0" mnemonicParsing="false" onAction="#submit" text="Submit" />
-                     </children>
-                  </AnchorPane>
-            </content>
-          </TitledPane>
-        </panes>
-      </Accordion>
-   </children>
-</Pane>
