+++ MissionControllerUI/src/main/resources/org/ariss/images/arissLogo.jpg
