+++ MissionControllerUI/src/main/resources/org/ariss/images/bluetooth.png
