+++ MissionControllerUI/src/main/resources/org/ariss/images/cancel.png
