+++ MissionControllerUI/src/main/resources/org/ariss/images/defaultMarsImage.jpg
