+++ MissionControllerUI/src/main/resources/org/ariss/images/recolor.png
