+++ MissionControllerUI/src/main/resources/org/ariss/images/robot.png
