+++ MissionControllerUI/src/main/resources/org/ariss/star/style.css
