+++ MissionControllerUI/src/main/resources/org/ariss/star/vis.css
